move operator contract logic to add parties page
